lihan 2020/6/5 20:54

+++ src/main/java/app/DetailBillItem.java
@@ -92,7 +92,7 @@ public class DetailBillItem {
         this.mode = mode;
     }
 
-    public double getDuration() {
+    public long getDuration() {
         return duration;
     }
